Apply branding to dealers app clip
