+++ Packages/EurofurenceComponents/Sources/ComponentBase/Branding/UIColor+Branding.swift
@@ -2,13 +2,92 @@ import UIKit
 
 extension UIColor {
     
-    public static let placeholder = adaptiveColor(lightColor: .pantone330U, darkColor: safeSystemGray2)
+    // MARK: Branding Colors
+    
+    public static let placeholder = adaptive(light: .pantone330U, dark: safeSystemGray2)
     public static let pantone330U = unsafelyNamed("Pantone 330U")
     public static let pantone330U_45 = unsafelyNamed("Pantone 330U (45%)")
     public static let pantone330U_26 = unsafelyNamed("Pantone 330U (26%)")
     public static let pantone330U_13 = unsafelyNamed("Pantone 330U (13%)")
     public static let pantone330U_5 = unsafelyNamed("Pantone 330U (5%)")
     public static let largeActionButton = unsafelyNamed("Large Action Button")
+    
+    // MARK: Semantic Control Colors
+    
+    public static let tintColor = adaptive(light: .pantone330U, dark: .pantone330U_45)
+    public static let disabledColor = safeSystemGray
+    public static let navigationBar = barColor
+    public static let tabBar = barColor
+    public static let searchBarTint = pantone330U
+    public static let refreshControl = pantone330U_13
+    public static let selectedTabBarItem = adaptive(light: .white, dark: .pantone330U_45)
+    public static let unselectedTabBarItem = adaptive(light: .pantone330U_45, dark: .darkGray)
+    public static let primary = adaptive(light: .pantone330U, dark: .black)
+    public static let secondary = adaptive(light: .pantone330U_45, dark: .secondaryDarkColor)
+    public static let buttons = pantone330U
+    public static let tableIndex = pantone330U
+    public static let iconographicTint = pantone330U
+    public static let unreadIndicator = pantone330U
+    public static let selectedSegmentText = adaptive(light: .pantone330U, dark: .white)
+    public static let selectedSegmentBackground = adaptive(light: .white, dark: .safeSystemGray)
+    public static let unselectedSegmentText = adaptive(light: .white, dark: .white)
+    public static let unselectedSegmentBackground = adaptive(light: .pantone330U_45, dark: .safeSystemGray3)
+    public static let segmentSeperator = adaptive(light: .white, dark: .safeSystemGray)
+    public static let safariBarTint = navigationBar
+    public static let safariControlTint = white
+    public static let userPrompt = adaptive(
+        light: UIColor(white: 0.5, alpha: 1.0),
+        dark: .safeSystemGray
+    )
+    
+    public static let userPromptWithUnreadMessages = pantone330U
+    
+    private static let barColor: UIColor = adaptive(light: .pantone330U, dark: calendarStyleBarColor)
+    
+    private static var calendarStyleBarColor: UIColor {
+        scaled(red: 18, green: 19, blue: 18)
+    }
+    
+    private static var secondaryDarkColor: UIColor = {
+        if #available(iOS 13.0, *) {
+            return .opaqueSeparator
+        } else {
+            return .black
+        }
+    }()
+    
+    private static var safeSystemGray: UIColor {
+        if #available(iOS 13.0, *) {
+            return .systemGray
+        } else {
+            return .lightGray
+        }
+    }
+    
+    private static var safeSystemGray3: UIColor {
+        if #available(iOS 13.0, *) {
+            return .systemGray3
+        } else {
+            return .darkGray
+        }
+    }
+    
+    private static func scaled(red: CGFloat, green: CGFloat, blue: CGFloat) -> UIColor {
+        let scale: (CGFloat) -> CGFloat = { $0 / 255.0 }
+        return UIColor(red: scale(red), green: scale(green), blue: scale(blue), alpha: 1.0)
+    }
+    
+    func makeColoredImage(size: CGSize) -> UIImage {
+        let renderer = UIGraphicsImageRenderer(size: size)
+        return renderer.image { (context) in
+            setFill()
+            context.fill(CGRect(origin: .zero, size: size))
+        }
+    }
+    
+    func makePixel() -> UIImage {
+        makeColoredImage(size: CGSize(width: 1, height: 1))
+    }
     
     private static var safeSystemGray2: UIColor {
         if #available(iOS 13.0, *) {
@@ -26,17 +105,17 @@ extension UIColor {
         return color
     }
     
-    public static func adaptiveColor(lightColor: UIColor, darkColor: UIColor) -> UIColor {
+    private static func adaptive(light: UIColor, dark: UIColor) -> UIColor {
         if #available(iOS 13.0, *) {
             return UIColor(dynamicProvider: { (traitCollection) in
                 if traitCollection.userInterfaceStyle == .light {
-                    return lightColor
+                    return light
                 } else {
-                    return darkColor
+                    return dark
                 }
             })
         } else {
-            return lightColor
+            return light
         }
     }
